requirements.txt
lark-parser
graphviz
pydot